[FEAT] Code Refactoring
